% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_obs.R
\name{between_obs}
\alias{between_obs}
\title{Check which assess a point lies between}
\usage{
between_obs(event_time, assess)
}
\arguments{
\item{event_time}{A vector of event times}

\item{assess}{A vector of assessment times to check. Not necessarily
in order}
}
\value{
A matrix of lower and upper bounds from \code{assess} that
each \code{event_time} lies between.
}
\description{
Check which assess a point lies between
}
\details{
This treats the intervals between observation times as open on
the bottom and closed on top. If \code{event_time} lies exactly on an
observation, that observation will be the upper.
If \code{event_time} doesn't lie between any of the assess, +/-
\code{Inf}
will be returned along with the closest observation.
}
\examples{
between_obs(0.5, c(0, 1, 2, 3, 4)) # returns 0, 1
between_obs(5, c(1, 2, 3, 4, 5, 6)) # returns 5, 5
between_obs(0, c(1, 2, 3, 4, 5, 6)) # returns -Inf, 1
between_obs(9, c(1, 2, 3, 4, 5, 6)) # returns 6, Inf
}
