% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_intervals.R
\name{inner_intervals}
\alias{inner_intervals}
\title{Find inner-intervals from observation times}
\usage{
inner_intervals(left, right, trunc = NULL, open_L = TRUE, open_T = FALSE)
}
\arguments{
\item{left}{Last assessment time before event}

\item{right}{First assessment time after event}

\item{trunc}{Truncation time}

\item{open_L}{Are left assessment times treated as open?}

\item{open_T}{Are truncation times treated as open?}
}
\value{
Data-frame containing left and right inner-intervals and
open/closed status of each end-point.
}
\description{
Find inner-intervals from observation times
}
\details{
The NPMLE for left-truncated and interval-censored data only changes across
a reduced number of inner-intervals (also known as Turnbull intervals or
maximal cliques). This function calculates these inner-intervals as suggested
by Yu (2022) to calculate the correct NPMLE. See Yu (2022) for details on
open or closed end-points.
}
\examples{
trunc <- c(0,   0, 0,   1,    1,   1)
left <- c(-Inf, 1, 2,   -Inf, 1,   2)
right <- c(1,   2, Inf, 1,    2, Inf)
inner_intervals(left, right, trunc)
ltic_np(left, right, trunc)
}
\references{
Yu, Qiqing. "The generalised MLE with truncated interval-censored data." Journal of Nonparametric Statistics 35.2 (2023): 266-282.
}
