% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai_ying.R
\name{lai_ying_plot}
\alias{lai_ying_plot}
\title{Plot Lai-Ying survival curve and risk-sets}
\usage{
lai_ying_plot(x, c = 1, alpha = 0.25, ...)
}
\arguments{
\item{x}{Output from \code{ltic_np()} using the default (PL-ICM) method}

\item{c}{Lai-Ying parameter}

\item{alpha}{Lai-Ying parameter}

\item{...}{Additional arguments for plotting}
}
\description{
Calculate and plot a Lai-Ying survival curve from the NPMLE for
left-truncated and interval
censored data. This is useful when, due to truncation, there are very small
risk-set sizes at early times which would usually result in underestimating
the survival curve. This is experimental because Lai-Ying applied their
method to left-truncated and right-censored data without interval-censoring.
}
\details{
The Lai-Ying estimator acts like the usual Kaplan-Meier estimator in the
presence of left-truncation, but ignores any events at times when the
risk-set is too small. Lai and Ying define this as a risk-set smaller than
\eqn{c n^{\alpha}}. We use their suggested values of
\eqn{c = 1} and \eqn{\alpha = 0.25} as defaults.

For left-truncated and interval-censored data, we use the expected risk-set
and event numbers and apply the Lai-Ying estimator as with left-truncated
and right-censored data.

Two plots are produced, one is a survival curve from the Lai-Ying estimator,
the other is the expected risk-set size and expected number of events from
the PL-ICM algorithm.
}
\examples{
est <- ltic_np(mhcps$Ui, mhcps$Vi, mhcps$Ti)
lai_ying_plot(est)
}
\references{
Lai, Tze Leung, and Zhiliang Ying. "Estimating a distribution function with truncated and censored data." The Annals of Statistics (1991): 417-442.
}
