% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_np.R
\name{lb_np}
\alias{lb_np}
\title{Calculate maximum likelihood of survival function}
\usage{
lb_np(
  left,
  right,
  tol = 1e-07,
  init = NULL,
  max_it = 1e+05,
  open_L = TRUE,
  open_T = FALSE,
  constr = NULL
)
}
\arguments{
\item{constr}{}
}
\description{
Calculate maximum likelihood of survival function
}
