% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_method.R
\name{optim_method}
\alias{optim_method}
\title{NPMLE for left-truncated interval-censored data using quasi-Newton method}
\usage{
optim_method(init, alpha, beta, tol = 1e-09)
}
\arguments{
\item{init}{Initial values for the probability masses in each inner-interval}

\item{alpha}{Indicator matrix of events in each inner-interval}

\item{beta}{Indicator matrix of truncation times for each inner-interval}

\item{tol}{Tolerance}
}
\description{
Wrapper for the \code{optim()} function to calculate the NPMLE
}
\details{
Using a method as roughly described by Hudgens (2005) (the exact
details aren't given there)
}
\references{
Hudgens, M. G. (2005). On nonparametric maximum likelihood estimation with interval censoring and left truncation. Journal of the Royal Statistical Society Series B: Statistical Methodology, 67(4), 573-587.
}
\author{
Zachary Waller
}
