% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ltic.R
\name{plot.ltic}
\alias{plot.ltic}
\title{Plot estimates from self-consistency algorithm}
\usage{
\method{plot}{ltic}(x, cond = NULL, plot_type = c("new", "over"), ends = c("r", "l"), ...)
}
\arguments{
\item{x}{ltic object returned from \code{ltic_np}}

\item{cond}{Index of inner-interval to condition survival on}

\item{plot_type}{Should a new plot be made (\code{"new"}) or should the
plot go over an existing one (\code{"over"})?}

\item{ends}{Choose between step-changes at right (\code{"r"}) or left
(\code{"l"}) end-points or inner-intervals.}

\item{...}{Other arguments supplied to plot function.}
}
\description{
Plot estimates from self-consistency algorithm
}
\examples{
est <- ltic_np(mhcps$Ui, mhcps$Vi, mhcps$Ti)
plot(est, cond = 1)

}
