% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{plot_estimate}
\alias{plot_estimate}
\title{Plot estimates from self-consistency algorithm}
\usage{
plot_estimate(
  estimate,
  plot_over = FALSE,
  plot_type = c("gaps", "filled"),
  xlim = NULL,
  cond = NULL,
  ...
)
}
\arguments{
\item{plot_type}{Either \code{"filled"} for Turnbull intervals filled with
line or \code{"gaps"} for Turnbull intervals left blank.}

\item{turnbulls}{Data frame containing Turnbull intervals}

\item{s}{Estimates of step changes from EM aglorithm}

\item{right}{Only necessary when \code{plot_type} is \code{"filled"} to
get a right upper limit for the plot. Note: Turnbull intervals are where we
don't know the value}
}
\description{
Plot estimates from self-consistency algorithm
}
