% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ltic.R
\name{print.ltic}
\alias{print.ltic}
\title{Print an NPMLE estimate}
\usage{
\method{print}{ltic}(x, ...)
}
\arguments{
\item{x}{Estimate from \code{ltic_np()}}

\item{...}{Other printing options}
}
\description{
Print an NPMLE estimate
}
\details{
Gives method used for calculating the NPMLE, details about the
inner-intervals, log-likelihood and number of iterations. For left-truncated
data it will also warn about inner-intervals that may cause a large drop in
survival estimate. See Hudgens (2005) and Waller et al. (2025) for
details.
}
\examples{
est <- ltic_np(mhcps$Ui, mhcps$Vi, mhcps$Ti)
print(est)

}
\references{
Hudgens, Michael G. "On nonparametric maximum likelihood estimation with interval censoring and left truncation." Journal of the Royal Statistical Society Series B: Statistical Methodology 67.4 (2005): 573-587.
Waller, Zachary, et al. "A fast and stable algorithm for calculating the non-parametric maximum likelihood estimate ofleft-truncated and interval-censored data." (2025).
}
