% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_intervals.R
\name{inner_intervals}
\alias{inner_intervals}
\title{Find inner-intervals from observation times}
\usage{
inner_intervals(left, right, trunc = NULL, open_L = TRUE, open_T = FALSE)
}
\arguments{
\item{left}{Left end point of interval}

\item{right}{Right end point of interval}

\item{trunc}{Truncation time}
}
\value{
Data-frame containing left and right inner-intervals and open/closed
status of each end-point.
}
\description{
Find inner-intervals from observation times
}
\examples{
trunc <- c(0,   0, 0,   1,    1,   1)
left <- c(-Inf, 1, 2,   -Inf, 1,   2)
right <- c(1,   2, Inf, 1,    2, Inf)
inner_intervals(left, right, trunc)
turnbull_est(left, right, trunc)
}
\references{
Qiqing Yu "The generalised MLE with truncated interval-censored data 2022"
}
