% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai_ying.R
\name{lai_ying}
\alias{lai_ying}
\title{Calculate the Lai-Ying estimator}
\usage{
lai_ying(x, c = 1, alpha = 0.25)
}
\arguments{
\item{x}{Output from \code{ltic_np()} using the default (PL-ICM) method}

\item{c}{Lai-Ying parameter}

\item{alpha}{Lai-Ying parameter}
}
\description{
Calculate a Lai-Ying survival curve from the NPMLE for left-truncated and
interval censored data. This is useful when, due to truncation, there are
very small risk-set sizes at early times which would usually result in
underestimating the survival curve. This is experimental because Lai-Ying
applied their method to left-truncated and right-censored data without
interval-censoring.
}
\examples{
est <- ltic_np(mhcps$Ui, mhcps$Vi, mhcps$Ti)
ly_est <- lai_ying(est)
plot(ly_est)
}
\references{
Lai, Tze Leung, and Zhiliang Ying. "Estimating a distribution function with truncated and censored data." The Annals of Statistics (1991): 417-442.
}
