% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltic_np.R
\name{ltic_np}
\alias{ltic_np}
\title{Calculate maximum likelihood of survival function}
\usage{
ltic_np(
  left,
  right,
  trunc = NULL,
  tol = 1e-07,
  init = NULL,
  max_iterations = 1e+05,
  remove_rcens = TRUE,
  open_L = TRUE,
  open_T = FALSE,
  constr = NULL,
  method = c("comb", "em", "newton")
)
}
\arguments{
\item{constr}{}
}
\description{
Calculate maximum likelihood of survival function
}
