% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltic_np.R
\name{ltic_np}
\alias{ltic_np}
\title{Calculate maximum likelihood of survival function}
\usage{
ltic_np(
  left,
  right,
  trunc = NULL,
  tol = 1e-07,
  init = NULL,
  max_it = 1e+05,
  remove_rcens = TRUE,
  open_L = TRUE,
  open_T = FALSE,
  constr = NULL,
  method = c("both", "turnbull", "shen", "yu", "breslow", "prodlim", "bres_comb",
    "binomial", "optim", "turn_comb", "yu_comb", "icm", "surv")
)
}
\arguments{
\item{constr}{}
}
\description{
Calculate maximum likelihood of survival function
}
