% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{plot.ltic}
\alias{plot.ltic}
\title{Plot estimates from self-consistency algorithm}
\usage{
\method{plot}{ltic}(estimate, cond = NULL, plot_type = "new", ends = "r", ...)
}
\arguments{
\item{estimate}{ltic object returned from \code{ltic_np}}

\item{cond}{index of inner-interval to condition survival on}

\item{...}{Other arguments supplied to plot function.}
}
\description{
Plot estimates from self-consistency algorithm
}
\examples{
data <- 

}
